Refactor EchoEvolver into package module
