+++ echo/__init__.py
@@ -1,5 +1,14 @@
 """Echo toolkit shared utilities."""
 
+from .evolver import EchoEvolver, EmotionalDrive, EvolverState, SystemMetrics, main
 from .thoughtlog import ThoughtLogger, thought_trace
 
-__all__ = ["ThoughtLogger", "thought_trace"]
+__all__ = [
+    "EchoEvolver",
+    "EmotionalDrive",
+    "EvolverState",
+    "SystemMetrics",
+    "ThoughtLogger",
+    "thought_trace",
+    "main",
+]
